Replacing Json config with QSettings
This commit replaces the Json-based configuration in `./config.json` with a pwd-independent `QSettings` setup and removes the now unused Json class.
The structure of the new `ConfigurationHandler` is about the same as the `Json` class, but replaces the Json-construction and file reading with `QSettings` calls.
Under *nix, the configuration can now be located under:
  `~/.config/BambooTracker/BambooTracker.ini`