Fix build on *BSD operating systems
When *BSD as operating system is detected:

* Use OSS (OpenSoundSystem) by default
* Accept use_alsa as qmake CONFIG option
* Accept use_pulse as qmake CONFIG option
* Accept use_jack, jack_has_rename as qmake CONFIG option

BambooTracker compiles with all the options but at this time only use_pulse is
verfied to work.

It has been discussed to keep the options in place to have the chance to
introduce runtime fixes at a later time.

Relax pedantic CPPFLAGS with -Wno-zero-length-array because FreeBSD ALSA headers
use zero-length array in /usr/local/include/alsa/pcm.h:597.