Fully fix bug where Fxx tempo effects slow down the song
This changes Fxx tempo effects to not reset the tempo state. In the
process, I changed TickCounter so calling resetRest() multiple times on
the same tick produces the same outcome as only calling it the last
time.

I also changed Fxx *speed* effects to not reset the tempo state (a
debatable choice), by changing setSpeed() to not set tickDiffSum_ to 0.
This is because in BambooTracker and FamiTracker, changing speed doesn't
cause previously set tempo to stop taking effect, so I decided to not
reset the tempo state either. However this has the downside that you
can't reset the tempo accumulator mid-song using a speed command.