Fix song hanging when playing a groove with values of 0
TickCounter::resetRest() computes the number of ticks to wait before the
next row and writes it to restTickToNextStep_. Prior to this commit, in
the case of a groove with values of 0, it wrote 0, and
TickCounter::countUp() decrements restTickToNextStep_, resulting in a
value of -1. Afterwards, TickCounter::countUp() has to run for 4 billion
ticks (plus signed integer overflow UB), decrementing it once per tick,
for it to reach 0 and for playback to advance to the next row.

To avoid this, clamp non-positive values to 1 in both speed and groove.