Round QPainter DPI scale factor to integer on Qt 5.14+
On Qt 5.14+, rounding 150% down to 1x painter scaling results in less
awkwardly large widget padding and non-pixelated icons. The downside is
that toolbar icons are smaller, making them harder to see and click.

On Qt 6, this ensures that a QWidget's size in physical pixels is a
integer multiple of its size in virtual pixels and remains stable upon
resizing, so we can create QPixmaps the same size as a widget which can
be painted 1:1 onto the widget. Additionally, QPainter operations at
integer virtual-pixel coordinates remain physical-pixel-aligned
regardless of the screen's DPI, fixing both custom painting and Qt's
Windows theme.

Text remains scaled to fractional DPI.