Fix instrument editor envelope graphs on Breeze master
Previously, FMOperatorTable constructed envFrame as a QFrame with a
sunken StyledPanel border, then used an event filter to draw onto
envFrame before it drew the StyledPanel border. This had an issue where
on Breeze master, the StyledPanel would then draw an opaque background
overwriting the graph. Additionally, on adwaita-qt, the QFrame's border
was 1 pixel smaller than the widget's bounding box (and had rounded
corners), and calling painter.eraseRect() produced a discolored
rectangle around the QFrame's border.

This hard-codes the graph to cover the entire rect() of the QFrame, then
allows the QFrame to cut off the edges. This preserves the prior graph
size (the QPixmap covered the entire envFrame), but allows the QFrame's
borders to clip the edges of the paint operation, and prevents the
QFrame from overwriting the paint with a solid background.

Using a QLayout to position envGraph in envFrame, or calling
ui->envGraph->setGeometry(ui->envFrame->contentsRect()), results in wide
gaps around the graph on some themes. The most noticeable gap is in the
adwaita-qt theme, with MacOS having a 1 pixel gap.

In general it's impossible to find the thickness of the frame.
QFrame::frameWidth() is wrong on adwaita-qt and says the frame is 4
pixels wide, despite the border line only covering up the outermost 2
pixels. (adwaita-qt hard-codes QAbstractScrollArea to report a narrower
2 pixel border.)