Switch instrument editor graph from QPixmap to QImage
This changes `envmap_` to a QImage with pixel format
`Format_ARGB32_Premultiplied`.

The envelope graph is drawn to `envmap_` before being blitted to the
QAbstractScrollArea's viewport. `envmap_` used to be a QPixmap, whose
pixel format is determined by the platform. On Windows, QPixmaps are
Format_RGB32, which doesn't support transparency. The problem is that
when drawing the image, we fill it with instFMEnvBackColor, which is
transparency, causing the image to be white instead of transparent.

This technically isn't an issue on Windows, because the
QAbstractScrollArea's viewport is white by default, and on Windows
BambooTracker doesn't support custom dark styles based on Fusion rather
than WindowsVista, and Qt doesn't support Windows 10's native dark
style. And on Linux, QPixmap is Format_ARGB32_Premultiplied, which
supports transparency properly. However, it's risky to depend on these
implementation details.

To ensure consistent correct behavior across platforms, this switches
`envmap_` to a QImage, which allows us to specify the exact pixel
format we want. And `Format_ARGB32_Premultiplied` is faster for alpha
compositing than `Format_ARGB32`, so we use premultiplied.